Refactor EmergencyGroup and GroupInvite services to use Long for ID parameters

+++ src/main/java/no/ntnu/stud/idatt2106/backend/controller/EmergencyGroupController.java
@@ -46,7 +46,7 @@ public class EmergencyGroupController {
 
   @Operation(summary = "Delete an emergency group by ID")
   @DeleteMapping("/{id}")
-  public ResponseEntity<Void> delete(@PathVariable int id) {
+  public ResponseEntity<Void> delete(@PathVariable Long id) {
     return service.delete(id) ? ResponseEntity.noContent().build()
         : ResponseEntity.notFound().build();
   }
@@ -60,7 +60,7 @@ public class EmergencyGroupController {
    */
   @Operation(summary = "Update an emergency group by ID")
   @PutMapping("/{id}")
-  public ResponseEntity<Void> update(@PathVariable int id,
+  public ResponseEntity<Void> update(@PathVariable Long id,
       @RequestBody EmergencyGroupRequest request) {
     return service.update(id, request)
         ? ResponseEntity.ok().build()
@@ -75,7 +75,7 @@ public class EmergencyGroupController {
    */
   @Operation(summary = "Get an emergency group by ID")
   @GetMapping("/{id}")
-  public ResponseEntity<EmergencyGroupResponse> getById(@PathVariable int id) {
+  public ResponseEntity<EmergencyGroupResponse> getById(@PathVariable Long id) {
     return repository.findById(id)
         .map(mapper::toResponse)
         .map(ResponseEntity::ok)
@@ -85,7 +85,7 @@ public class EmergencyGroupController {
   @Operation(summary = "Get emergency group summaries for a household")
   @GetMapping("/summary/{householdId}")
   public ResponseEntity<List<EmergencyGroupSummaryResponse>> getGroupSummaries(
-      @PathVariable int householdId) {
+      @PathVariable Long householdId) {
     return ResponseEntity.ok(service.getGroupSummariesByHouseholdId(householdId));
   }
